add space barista

+++ backend_coffee_shop_php/routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-
+use App\Enum\UserRole;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\web\AuthController;
 use App\Http\Controllers\web\HomeController;
@@ -12,22 +12,25 @@ use App\Http\Controllers\web\ProductsController as ProductControllerWeb;
 use App\Http\Controllers\web\SettingController;
 use App\Http\Controllers\web\StockLogController;
 
-Route::get('/', HomeController::class)->name('home')->middleware('auth');
+$admin = UserRole::ADMIN->value;
+$barista = UserRole::BARISTA->value;
 
+
+Route::get('/', HomeController::class)->name('home')->middleware(["auth"]);
 
 Route::controller(UsersController::class)
 
     ->prefix('users')
     ->name('users.')
-    ->middleware('auth')
-    ->group(function () {
+    ->middleware(["auth"])
+    ->group(function () use ($admin) {
 
         Route::get('/', 'index')->name('index');
-        Route::post('/',  'store')->name('store');
-        Route::put('/{id}',  'update')->name('update');
-        Route::delete('/{id}',  'destroy')->name('destroy');
-        Route::get('/{id}/edit',  'edit')->name('edit');
-        Route::patch('/{id}/toggle-blocked',  'toggleBlocked')->name('toggleBlocked');
+        Route::post('/',  'store')->middleware("role:$admin")->name('store');
+        Route::put('/{id}',  'update')->middleware("role:$admin")->name('update');
+        Route::delete('/{id}',  'destroy')->middleware("role:$admin")->name('destroy');
+        Route::get('/{id}/edit',  'edit')->middleware("role:$admin")->name('edit');
+        Route::patch('/{id}/toggle-blocked',  'toggleBlocked')->middleware("role:$admin")->name('toggleBlocked');
     });
 
 
@@ -35,15 +38,15 @@ Route::controller(ProductControllerWeb::class)
 
     ->prefix('products')
     ->name('products.')
-    ->middleware('auth')
-    ->group(function () {
+    ->middleware(["auth"])
+    ->group(function () use ($admin) {
 
         Route::get('/', 'index')->name('index');
-        Route::post('/', 'store')->name('store');
-        Route::get('/{id}/edit', 'edit')->name('edit');
-        Route::delete('/{id}', 'destroy')->name('destroy');
-        Route::put('/{id}', 'update')->name('update');
-        Route::patch('/{id}/toggle-visibility', 'toggleVisibility')->name('toggleVisibility');
+        Route::post('/', 'store')->middleware("role:$admin")->name('store');
+        Route::get('/{id}/edit', 'edit')->middleware("role:$admin")->name('edit');
+        Route::delete('/{id}', 'destroy')->middleware("role:$admin")->name('destroy');
+        Route::put('/{id}', 'update')->middleware("role:$admin")->name('update');
+        Route::patch('/{id}/toggle-visibility', 'toggleVisibility')->middleware("role:$admin")->name('toggleVisibility');
     });
 
 
@@ -51,26 +54,26 @@ Route::controller(ProductControllerWeb::class)
 Route::controller(OrdersController::class)
     ->prefix('orders')
     ->name('orders.')
-    ->middleware('auth')
+    ->middleware(["auth"])
 
-    ->group(function () {
+    ->group(function () use ($admin) {
 
         Route::get('/', 'index')->name('index');
-        Route::get('/{id}', 'show')->name('show');
-        Route::delete('/{id}', 'destroy')->name('destroy');
+        Route::get('/{id}', 'show')->middleware("role:$admin")->name('show');
+        Route::delete('/{id}', 'destroy')->middleware("role:$admin")->name('destroy');
     });
 
 
 Route::controller(CategoryController::class)
     ->prefix('categories')
     ->name('categories.')
-    ->middleware('auth')
-    ->group(function () {
+    ->middleware(["auth"])
+    ->group(function () use ($admin) {
         Route::get('/', 'index')->name('index');
-        Route::post('/', 'store')->name('store');
-        Route::delete('/{category}', 'destroy')->name('destroy');
-        Route::put('/{category}', 'update')->name('update');
-        Route::get('/{category}/edit', 'edit')->name('edit');
+        Route::post('/', 'store')->middleware("role:$admin")->name('store');
+        Route::delete('/{category}', 'destroy')->middleware("role:$admin")->name('destroy');
+        Route::put('/{category}', 'update')->middleware("role:$admin")->name('update');
+        Route::get('/{category}/edit', 'edit')->middleware("role:$admin")->name('edit');
     });
 
 
@@ -86,13 +89,13 @@ Route::controller(AuthController::class)->group(function () {
 Route::controller(IngredientController::class)
     ->prefix('ingredients')
     ->name('ingredients.')
-    ->middleware('auth')
-    ->group(function () {
+    ->middleware(["auth"])
+    ->group(function () use ($admin) {
         route::get('/', 'index')->name('index');
-        route::post('/', 'store')->name('store');
-        route::get('/{ingredient}/edit', 'edit')->name('edit');
-        route::delete('/{ingredient}', 'destroy')->name('destroy');
-        route::put('/{ingredient}', 'update')->name('update');
+        route::post('/', 'store')->middleware("role:$admin")->name('store');
+        route::get('/{ingredient}/edit', 'edit')->middleware("role:$admin")->name('edit');
+        route::delete('/{ingredient}', 'destroy')->middleware("role:$admin")->name('destroy');
+        route::put('/{ingredient}', 'update')->middleware("role:$admin")->name('update');
     });
 
 
@@ -109,11 +112,9 @@ Route::controller(StockLogController::class)
 Route::controller(SettingController::class)
     ->prefix('setting')
     ->name('setting.')
-    ->middleware('auth')
-    ->group(function () {
-
+    ->middleware(["auth"])
+    ->group(function () use ($admin) {
         Route::get('/', 'index')->name('index');
         Route::get('/lang/{locale}', 'lang')->name('lang');
-        Route::put('/', 'update')->name('update');
-
+        Route::put('/', 'update')->middleware("role:$admin")->name('update');
     });;
